requirements.txt
Flask==2.0.2
fpdf==1.7.2
openpyxl==3.0.9
pyshp==2.1.3
python_dateutil==2.8.2
treetopper==1.1.6
xlrd==2.0.1
